refactor(v1.2.6): split security.py into 11 atomic SRP modules + facade
Decomposes security.py (169 lines, 4 auth schemes + 2 credential types + parser
in one file) into 11 atomic modules. One file per auth concept.

New structure (tachyon_api/security/):
  - _bearer_credentials.py   — HTTPAuthorizationCredentials value object
  - _basic_credentials.py    — HTTPBasicCredentials value object
  - _bearer_parser.py        — parse_bearer_header() pure function
  - _http_bearer.py          — HTTPBearer scheme
  - _http_basic.py           — HTTPBasic scheme (base64 + WWW-Authenticate)
  - _api_key_base.py         — _APIKeyBase ABC
  - _api_key_header.py       — APIKeyHeader
  - _api_key_query.py        — APIKeyQuery (security warning docstring)
  - _api_key_cookie.py       — APIKeyCookie
  - _oauth2_bearer.py        — OAuth2PasswordBearer
  - __init__.py              — re-exports the 8 public symbols

Public API unchanged: every symbol previously importable from
tachyon_api.security remains accessible at the same path. All 360/361 tests pass.

Refinements:
- All classes gain __slots__ declarations.
- _APIKeyBase uses @abstractmethod (was: raises NotImplementedError) — caught
  at class-definition time instead of runtime.
- HTTPBasic's reject-with-401 logic factored into _reject() helper.
- _extract_bearer renamed to parse_bearer_header (more descriptive, cdef-ready).

Bumps version to 1.2.6.
